requirements.txt
wxpython
pywin32
markdown
vk_api
vkaudiotoken
bs4
configobj
pypubsub
requests-oauthlib
future
arrow
backports.functools_lru_cache
googletrans
mutagen
mock
babel
nuitka
pyenchant
git+https://github.com/accessibleapps/libloader
git+https://github.com/accessibleapps/platform_utils
git+https://github.com/accessibleapps/sound_lib
git+https://github.com/accessibleapps/accessible_output2